requirements.txt
PyYAML
beautifulsoup4
lxml
selenium